Add WIP new bdf parser
The goal with this parser is to better group related code and have
the parsers own the data its parsing to make it easier to keep parsing
state together with the data to parse. Its far from finished, though
it can borrow heavily from the previous parser.

I put it in the master branch under v1 rather than a different branch
cause I would like to keep all code functional as its being worked on
rather than worry about changes on something else breaking this code, or
causing a difficult merge.  The goal would be to eventually replace the
v0 parser with this one and eliminate the duplicate code